# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BSV blockchain indexer API for transaction outputs, blocks, and related data",
        "title": "1Sat Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/shruggr/1sat-indexer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "5.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v5/origins/ancestors": {
            "post": {
                "description": "Get ancestors for multiple origins by outpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "origins"
                ],
                "summary": "Get ancestors for multiple origins",
                "parameters": [
                    {
                        "description": "Array of origin outpoints",
                        "name": "outpoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/origins/ancestors/{outpoint}": {
            "get": {
                "description": "Get ancestors of a specific origin (excluding the origin itself)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "origins"
                ],
                "summary": "Get origin ancestors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Origin outpoint",
                        "name": "outpoint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/origins/history": {
            "post": {
                "description": "Get the history for multiple origins by outpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "origins"
                ],
                "summary": "Get multiple origins history",
                "parameters": [
                    {
                        "description": "Array of origin outpoints",
                        "name": "outpoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/origins/history/{outpoint}": {
            "get": {
                "description": "Get the history of a specific origin by outpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "origins"
                ],
                "summary": "Get origin history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Origin outpoint",
                        "name": "outpoint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "evt.Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "idx.IndexData": {
            "type": "object",
            "properties": {
                "data": {},
                "deps": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/evt.Event"
                    }
                }
            }
        },
        "idx.Txo": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/idx.IndexData"
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "outpoint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "satoshis": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "script": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spend": {
                    "type": "string"
                }
            }
        }
    }
}