# server/routes/origins/ctrl.go
package origins

import (
	"encoding/json"
	"strings"

	"github.com/gofiber/fiber/v2"
	"github.com/shruggr/1sat-indexer/v5/evt"
	"github.com/shruggr/1sat-indexer/v5/idx"
	"github.com/shruggr/1sat-indexer/v5/mod/onesat"
)

var ingest *idx.IngestCtx

func RegisterRoutes(r fiber.Router, ingestCtx *idx.IngestCtx) {
	ingest = ingestCtx
	r.Post("/ancestors", OriginsAncestors)
	r.Get("/ancestors/:outpoint", OriginAncestors)
	r.Post("/history", OriginsHistory)
	r.Get("/history/:outpoint", OriginsHistory)
}

// @Summary Get origin history
// @Description Get the history of a specific origin by outpoint
// @Tags origins
// @Produce json
// @Param outpoint path string true "Origin outpoint"
// @Param tags query string false "Comma-separated list of tags to include (use * for all indexed tags)"
// @Param txo query bool false "Include TXO data"
// @Param script query bool false "Include script data"
// @Success 200 {array} idx.Txo
// @Failure 500 {string} string "Internal server error"
// @Router /v5/origins/history/{outpoint} [get]
func OriginHistory(c *fiber.Ctx) error {
	outpoint := c.Params("outpoint")
	tags := strings.Split(c.Query("tags", ""), ",")
	if len(tags) > 0 && tags[0] == "*" {
		tags = ingest.IndexedTags()
	}

	if txos, err := ingest.Store.SearchTxos(c.Context(), &idx.SearchCfg{
		Keys: []string{evt.EventKey("origin", &evt.Event{
			Id:    "outpoint",
			Value: outpoint,
		})},
		IncludeTxo:    c.QueryBool("txo", false),
		IncludeTags:   tags,
		IncludeScript: c.QueryBool("script", false),
	}); err != nil {
		return err
	} else {
		return c.JSON(txos)
	}

}

// @Summary Get multiple origins history
// @Description Get the history for multiple origins by outpoints
// @Tags origins
// @Accept json
// @Produce json
// @Param outpoints body []string true "Array of origin outpoints"
// @Param tags query string false "Comma-separated list of tags to include (use * for all indexed tags)"
// @Param txo query bool false "Include TXO data"
// @Param script query bool false "Include script data"
// @Success 200 {array} idx.Txo
// @Failure 400 {string} string "Invalid request"
// @Failure 500 {string} string "Internal server error"
// @Router /v5/origins/history [post]
func OriginsHistory(c *fiber.Ctx) error {
	var outpoints []string
	if err := c.BodyParser(&outpoints); err != nil {
		return c.SendStatus(400)
	} else if len(outpoints) == 0 {
		return c.SendStatus(400)
	}

	tags := strings.Split(c.Query("tags", ""), ",")
	if len(tags) > 0 && tags[0] == "*" {
		tags = ingest.IndexedTags()
	}

	history := make([]*idx.Txo, 0, len(outpoints))
	for _, outpoint := range outpoints {
		if txos, err := ingest.Store.SearchTxos(c.Context(), &idx.SearchCfg{
			Keys: []string{evt.EventKey("origin", &evt.Event{
				Id:    "outpoint",
				Value: outpoint,
			})},
			IncludeTxo:    c.QueryBool("txo", false),
			IncludeTags:   tags,
			IncludeScript: c.QueryBool("script", false),
		}); err != nil {
			return err
		} else {
			history = append(history, txos...)
		}
	}
	return c.JSON(history)
}

// @Summary Get origin ancestors
// @Description Get ancestors of a specific origin (excluding the origin itself)
// @Tags origins
// @Produce json
// @Param outpoint path string true "Origin outpoint"
// @Param tags query string false "Comma-separated list of tags to include (use * for all indexed tags)"
// @Param txo query bool false "Include TXO data"
// @Param script query bool false "Include script data"
// @Success 200 {array} idx.Txo
// @Failure 500 {string} string "Internal server error"
// @Router /v5/origins/ancestors/{outpoint} [get]
func OriginAncestors(c *fiber.Ctx) error {
	outpoint := c.Params("outpoint")

	tags := strings.Split(c.Query("tags", ""), ",")
	if len(tags) > 0 && tags[0] == "*" {
		tags = ingest.IndexedTags()
	}

	if data, err := ingest.Store.LoadData(c.Context(), outpoint, []string{"origin"}); err != nil {
		return err
	} else if originData, ok := data["origin"]; ok && originData != nil {
		origin := onesat.Origin{}
		if err := json.Unmarshal([]byte(originData.Data.(json.RawMessage)), &origin); err != nil {
			return err
		}
		outpoint = origin.Outpoint.String()
	}

	ancestors := make([]*idx.Txo, 0)
	if txos, err := ingest.Store.SearchTxos(c.Context(), &idx.SearchCfg{
		Keys: []string{evt.EventKey("origin", &evt.Event{
			Id:    "outpoint",
			Value: outpoint,
		})},
		IncludeTxo:    c.QueryBool("txo", false),
		IncludeTags:   tags,
		IncludeScript: c.QueryBool("script", false),
	}); err != nil {
		return err
	} else {
		for _, txo := range txos {
			if txo.Outpoint.String() != outpoint {
				ancestors = append(ancestors, txo)
			}
		}
	}
	return c.JSON(ancestors)
}

// @Summary Get ancestors for multiple origins
// @Description Get ancestors for multiple origins by outpoints
// @Tags origins
// @Accept json
// @Produce json
// @Param outpoints body []string true "Array of origin outpoints"
// @Param tags query string false "Comma-separated list of tags to include (use * for all indexed tags)"
// @Param txo query bool false "Include TXO data"
// @Param script query bool false "Include script data"
// @Success 200 {array} idx.Txo
// @Failure 400 {string} string "Invalid request"
// @Failure 500 {string} string "Internal server error"
// @Router /v5/origins/ancestors [post]
func OriginsAncestors(c *fiber.Ctx) error {
	var outpoints []string
	if err := c.BodyParser(&outpoints); err != nil {
		return c.SendStatus(400)
	} else if len(outpoints) == 0 {
		return c.SendStatus(400)
	}

	tags := strings.Split(c.Query("tags", ""), ",")
	if len(tags) > 0 && tags[0] == "*" {
		tags = ingest.IndexedTags()
	}

	origins := make([]string, 0, len(outpoints))
	outpointMap := make(map[string]struct{}, len(outpoints))
	for _, outpoint := range outpoints {
		if data, err := ingest.Store.LoadData(c.Context(), outpoint, []string{"origin"}); err != nil {
			return err
		} else if originData, ok := data["origin"]; ok && originData != nil {
			origin := onesat.Origin{}
			if err := json.Unmarshal([]byte(originData.Data.(json.RawMessage)), &origin); err != nil {
				return err
			}
			if origin.Outpoint != nil {
				origins = append(origins, origin.Outpoint.String())
			}
		}
	}

	ancestors := make([]*idx.Txo, 0, len(origins))
	for _, outpoint := range origins {
		if txos, err := ingest.Store.SearchTxos(c.Context(), &idx.SearchCfg{
			Keys: []string{evt.EventKey("origin", &evt.Event{
				Id:    "outpoint",
				Value: outpoint,
			})},
			IncludeTxo:    c.QueryBool("txo", false),
			IncludeTags:   tags,
			IncludeScript: c.QueryBool("script", false),
		}); err != nil {
			return err
		} else {
			for _, txo := range txos {
				op := txo.Outpoint.String()
				if _, ok := outpointMap[op]; !ok {
					outpointMap[op] = struct{}{}
					ancestors = append(ancestors, txo)
				}
			}
		}
	}

	return c.JSON(ancestors)
}